+++ doc/changelog.rst
@@ -1,8 +1,20 @@
 Changelog
 =========
 
+v0.1.1
+------
+*(unreleased)*
+
+- Fixed bug in :func:`~mizani.bounds.censor` where a sequence of
+  values with an irregular index would lead to an exception.
+
+
+
 v0.1.0
 ------
 *(2016-06-30)*
+
+.. image:: https://zenodo.org/badge/doi/10.5281/zenodo.57030.svg
+   :target: http://dx.doi.org/10.5281/zenodo.57030
 
 First public release